clean bugs

+++ js/controllers/controller.business.js
@@ -21,55 +21,56 @@ iCloudController.controller("ShopManagementController", ["$scope", "$http", "$gr
             $window.location.href = ["#/main/edit_shop?shop_id=", id].join("");
         };
         $scope.delete_shop = function (id) {
-            $http.delete([window.API.GROUP.REMOVE_GROUP, "?key=", $cookieStore.get("key"), "&id=", id].join("")).success(function (data) {
-                alert(data.msg);
-                show_shop();
-            })
+            if (confirm("确定删除?")) {
+                $http.delete([window.API.GROUP.REMOVE_GROUP, "?key=", $cookieStore.get("key"), "&id=", id].join("")).success(function (data) {
+                    show_shop();
+                })
+            }
         }
     }]);
 
-iCloudController.controller("CreateShopController", ["$scope", "$http", "$category", "$province", "$city", "$area", "$trades", "$cookieStore", function ($scope, $http, $category, $province, $city, $area, $trades, $cookieStore) {
-    $scope.shop = {};
-    $category.get().success(function (data) {
-        $scope.shop.category = data[0].name;
-        $scope.category = data;
-    })
-        .error(function (data) {
-            console.log(data);
-        });
-    $province.get().success(function (data) {
-        $scope.province = data;
-    });
-    $scope.select_p = function (id) {
-        $city.get(id).success(function (data) {
-            $scope.city = data;
-        })
-    };
-    $scope.select_c = function (id) {
-        $area.get(id).success(function (data) {
-            $scope.area = data;
-        })
-    }
-    $scope.select_a = function (id) {
-        $trades.get(id).success(function (data) {
-            $scope.trades = data;
-        })
-    };
-    $scope.submit = function (shop) {
-
-        $http.post([window.API.GROUP.NEW_GROUP, "?key=", $cookieStore.get("key")].join(""), shop).success(function (data) {
-            alert(data.msg);
-            location.href = "#/main/shop_management";
+iCloudController.controller("CreateShopController", ["$scope", "$http", "$category", "$province", "$city", "$area", "$trades", "$cookieStore", "$window",
+    function ($scope, $http, $category, $province, $city, $area, $trades, $cookieStore, $window) {
+        $scope.shop = {};
+        $category.get().success(function (data) {
+            $scope.shop.category = data[0].name;
+            $scope.category = data;
         })
             .error(function (data) {
                 console.log(data);
+            });
+        $province.get().success(function (data) {
+            $scope.province = data;
+        });
+        $scope.select_p = function (id) {
+            $city.get(id).success(function (data) {
+                $scope.city = data;
             })
-    }
-}]);
+        };
+        $scope.select_c = function (id) {
+            $area.get(id).success(function (data) {
+                $scope.area = data;
+            })
+        }
+        $scope.select_a = function (id) {
+            $trades.get(id).success(function (data) {
+                $scope.trades = data;
+            })
+        };
+        $scope.submit = function (shop) {
+
+            $http.post([window.API.GROUP.NEW_GROUP, "?key=", $cookieStore.get("key")].join(""), shop)
+                .success(function (data) {
+                    $window.location.href = "#/main/shop_management";
+                })
+                .error(function (data) {
+                    $window.alert("创建失败");
+                })
+        }
+    }]);
 
 iCloudController.controller("EditShopController", ["$scope", "$http", "$category", "$province", "$city", "$area", "$trades", "$cookieStore", function ($scope, $http, $category, $province, $city, $area, $trades, $cookieStore) {
     var id = get_param(window.location.href);
-    $scope.edit_shop;
     $http.get([window.API.GROUP.GET_CURRENT_USER_ROUTER_GROUPS, "?key=", $cookieStore.get("key")].join("")).success(function (data) {
         for (var i = 0; i < data.count; i++) {
             if (data.results[i].id == id) {
@@ -98,7 +99,7 @@ iCloudController.controller("EditShopController", ["$scope", "$http", "$category
         $area.get(id).success(function (data) {
             $scope.area = data;
         })
-    }
+    };
     $scope.select_a = function (id) {
         $trades.get(id).success(function (data) {
             $scope.trades = data;
@@ -106,10 +107,11 @@ iCloudController.controller("EditShopController", ["$scope", "$http", "$category
     };
     $scope.submit = function (shop) {
 
-        $http.put([window.API.GROUP.EDIT_GROUP, "?key=", $cookieStore.get("key")].join(""), shop).success(function (data) {
-            alert(data.msg);
-            location.href = "#/main/shop_management";
-        })
+        $http.put([window.API.GROUP.EDIT_GROUP, "?key=", $cookieStore.get("key")].join(""), shop)
+            .success(function (data) {
+                alert(data.msg);
+                location.href = "#/main/shop_management";
+            })
             .error(function (data) {
                 console.log(data);
             })
@@ -245,32 +247,32 @@ iCloudController.controller("RoutersDetailsController", ["$scope", "$http", "$co
         var routerStatusInterval = $interval(function () {
             $http.get([window.API.WIFICAT.STATUS, "?key=", $cookieStore.get("key"), "&router_mac=", data.router.mac].join("")).success(function (data) {
                 console.log(data.msg);
-                if(data.msg=="Router offline"){
+                if (data.msg == "Router offline") {
                     $scope.wificat = {
-                        "operatingStatus":{
-                            "accessNumber":"未连接",
-                            "MemUsaged":"未连接",
-                            "cpuUtil":"未连接"
+                        "operatingStatus": {
+                            "accessNumber": "未连接",
+                            "MemUsaged": "未连接",
+                            "cpuUtil": "未连接"
                         },
-                        "basicInformation":{
-                            "softwareVersion":"未连接"
+                        "basicInformation": {
+                            "softwareVersion": "未连接"
                         },
-                        "wanStatus":{
-                            "wanip":"未连接",
-                            "speedUp":"未连接",
-                            "speedDown":"未连接"
+                        "wanStatus": {
+                            "wanip": "未连接",
+                            "speedUp": "未连接",
+                            "speedDown": "未连接"
                         }
                     };
                     $scope.upTime = "未连接"
                 }
-                else{
+                else {
                     $scope.wificat = data;
                     $scope.upTime = parseInt(data.basicInformation.upTime / 60);
                 }
             });
         }, 1000);
         $scope.$on("$destroy", function () {
-           $interval.cancel(routerStatusInterval);
+            $interval.cancel(routerStatusInterval);
         });
 
         //默认认证方式
